add room templates finish with tests
