+++ src/main/java/missions/room/Service/AddRoomTemplateService.java
@@ -1,4 +1,4 @@
-package Service;
+package missions.room.Service;
 
 import DataAPI.Auth;
 import DataAPI.Response;
