+++ src/test/java/missions/room/AddMissionTests/AddMissionTestsRealRamMissionTeacherRepo.java
@@ -6,41 +6,22 @@ import RepositoryMocks.MissionRepository.MissionCrudRepositoryMock;
 import missions.room.Domain.Mission;
 import missions.room.Domain.Ram;
 import missions.room.Managers.AddMissionManager;
-import missions.room.Repo.MissionRepo;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
-
-import java.lang.reflect.Field;
-
-import static org.junit.jupiter.api.Assertions.fail;
+import org.springframework.test.context.TestPropertySource;
 
 @SpringBootTest
+@TestPropertySource(locations = {"classpath:application-unit-integration-tests.properties"})
 public class AddMissionTestsRealRamMissionTeacherRepo extends AddMissionTestsRealRamMission {
 
     @Override
     void setUpMocks() {
         missionCrudRepository=new MissionCrudRepositoryMock(dataGenerator);
         missionString=gson.toJson(dataGenerator.getMission(Data.Valid_Deterministic), Mission.class);
-        try {
-            Field missionRepo = AddMissionManager.class.getDeclaredField("missionRepo");
-            missionRepo.setAccessible(true);
-            missionRepo.set(addMissionManager,new MissionRepo(missionCrudRepository));
-        } catch (IllegalAccessException | NoSuchFieldException e) {
-            fail();
-        }
+        ram=new Ram();
+        addMissionManager=new AddMissionManager(ram,teacherCrudRepository,missionCrudRepository);
     }
 
-    @Override
-    void setUpAddMission(){
-        try {
-            Field managerRam = AddMissionManager.class.getDeclaredField("ram");
-            managerRam.setAccessible(true);
-            ram=(Ram)managerRam.get(addMissionManager);
-        } catch (IllegalAccessException | NoSuchFieldException e) {
-            fail();
-        }
-        super.setUpAddMission();
-    }
 
     @Test
     void testMissionInvalidWrongTeacher(){
