+++ src/test/java/missions/room/addRoomTemplateTests/AddRoomTemplateTestsRealRamTeacherRepoMissionRepo.java
@@ -9,12 +9,14 @@ import missions.room.Repo.RoomTemplateRepo;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.TestPropertySource;
 
 import java.lang.reflect.Field;
 
 import static org.junit.jupiter.api.Assertions.fail;
 
 @SpringBootTest
+@TestPropertySource(locations = {"classpath:application-unit-integration-tests.properties"})
 public class AddRoomTemplateTestsRealRamTeacherRepoMissionRepo extends AddRoomTemplateTestsRealRamTeacherRepo {
