+++ src/test/resources/application-unit-integration-tests.properties
@@ -1,6 +1,4 @@
-#spring.datasource.url=jdbc:h2:file:C:/Users/yuvalhuck2/IdeaProjects/MissionsRooms
-#;INIT=RUNSCRIPT FROM 'src/main/resources/init.sql';
-spring.datasource.url=jdbc:h2:mem:testdb
+#spring.datasource.url=jdbc:h2:mem:testdb;INIT=RUNSCRIPT FROM 'src/test/resources/init.sql';
 spring.jpa.hibernate.ddl-auto=update
 spring.datasource.driverClassName=org.h2.Driver
 spring.datasource.username=sa
@@ -8,5 +6,3 @@ spring.datasource.password=
 spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
 spring.h2.console.enabled=true
 #spring.datasource.initialize=true
-
-
